\name{snn-package}
\alias{snn-package}
\alias{snn}
\docType{package}
\title{
Package for Stabilized Nearest Neighbor Classifier
}
\description{
A package for implementations of various nearest neighbor classifiers, including K-nearest neighbor classifier, weighted nearest neighbor classifier, bagged nearest neighbor classifier, optimal weighted nearest neighbor classifier, and a new stabilized nearest neighbor classifier. This package also provides functions for computing the classification error and classification instability of a classification procedure.
}
\details{
\tabular{ll}{
Package: \tab snn\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-31\cr
License: \tab GPL-3 \cr
}
The package "snn" provides 8 main functions:
(1) the classification error.
(2) the classification instability.
(3) the K-nearest neighbor classifier.
(4) the weighted neighbor classifier.
(5) the bagged nearest neighbor classifier.
(6) the optimal nearest neighbor classifier.
(7) the stabilized nearest neighbor classifier.
(8) the model selection via cross-validation for K-nearest neighbor classifier, bagged nearest neighbor classifier, optimal nearest neighbor classifier, and stabilized nearest neighbor classifier.
}
\author{
Wei Sun, Xingye Qiao, and Guang Cheng

Maintainer: Wei Sun  <sunweisurrey8@gmail.com>
}
\references{
W. Sun, X. Qiao, and G. Cheng (2015) Stabilized Nearest Neighbor Classifier and Its Statistical Properties. Available at arxiv.org/abs/1405.6642. 
}
